replace person activity with fragment
